backend/push-notifications/requirements.txt
psycopg2-binary
pywebpush>=2.0.0
